+++ gb-breakpoint/src/native.rs
@@ -21,3 +21,11 @@ pub fn value(input: &str) -> IResult<&str, u16> {
         u16::from_str_radix(s, 16).unwrap()
     })(input)
 }
+
+#[test]
+fn test_raw_value() {
+    assert_eq!(value("1"), Ok(("", 0x1_u16)));
+    assert_eq!(value("1f"), Ok(("", 0x1f_u16)));
+    assert_eq!(value("b1f"), Ok(("", 0xb1f_u16)));
+    assert_eq!(value("ab1f"), Ok(("", 0xab1f_u16)));
+}
